# src/components/facility/facility.types.ts
export interface IFacility {
    name: string;
    address: string;
    description: string;
}

export interface IFacilityDTO {
    id: string,
    name: string,
    address: string,
    description: string,
}