# src/components/position/position.types.ts
export interface IPosition {
    name: string;
    description: string;
}

export interface IPositionDTO {
    id: string,
    name: string,
    description: string,
}

// Compare this snippet from src\components\Position\Position.controller.ts: